+++ packages/cpp/dwarf-therapist/default.nix
@@ -1,13 +1,13 @@
 { lib, stdenv, fetchFromGitHub, libsForQt5, cmake }:
 stdenv.mkDerivation rec {
   pname = "dwarf-therapist";
-  version = "42.1.2";
+  version = "42.1.5";
 
   src = fetchFromGitHub {
     owner = "Dwarf-Therapist";
     repo = "Dwarf-Therapist";
     rev = "v${version}";
-    sha256 = "sha256-gBYMYPW/JCHrlKF9SwXlCwjcDgC5LTYfrxlGdL1HXP8=";
+    sha256 = "sha256-aUakfUjnIZWNDhCkG3A6u7BaaCG8kPMV/Fu2S73CoDg=";
   };
 
   nativeBuildInputs = [ cmake ];